Experiment with prerendering

+++ web/scripts/build.js
@@ -46,9 +46,11 @@ measureFileSizesBeforeBuild(paths.appBuild)
   .then(previousFileSizes => {
     // Remove all content but keep the directory so that
     // if you're in it, you don't end up in Trash
-    fs.emptyDirSync(paths.appBuild);
-    // Merge with the public folder
-    copyPublicFolder();
+    if (process.env.PRERENDER === 'true') {
+      fs.emptyDirSync(paths.appBuild);
+      // Merge with the public folder
+      copyPublicFolder();
+    }
     // Start the webpack build
     return build(previousFileSizes);
   })
